WIP - Refactor/Consolidate input config to single file
